[#3] refactor: apply eslint for each files
