+++ src/main.ts
@@ -1,7 +1,7 @@
-import { createApp } from "vue";
-import App from "./App.vue";
-import router from "./router";
+import { createApp } from 'vue';
+import App from './App.vue';
+import router from './router';
 
 createApp(App)
-  .use(router)
-  .mount("#app");
+    .use(router)
+    .mount('#app');
